% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{get_distances}
\alias{get_distances}
\title{Implements a basic k-means algorithm.}
\usage{
get_distances(x, y, start_x, start_y)
}
\arguments{
\item{dat}{A data frame}

\item{start_points}{A data frame}
}
\value{
A data frame containing the distances of each point in dat to the initial points
}
\description{
Implements a basic k-means algorithm.
}
