% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{k_means}
\alias{k_means}
\title{Implements a basic k-means algorithm.}
\usage{
k_means(dat, k, pca = FALSE, max_iter = 1000)
}
\arguments{
\item{dat}{A data frame}

\item{k}{A number}

\item{pca}{A boolean}
}
\value{
A list with a dataframe containing the original x & y values of the input data and a factor which contains the grouping variable, and the total sum of squares
}
\description{
Implements a basic k-means algorithm.
}
